Avoid two warnings
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@11968 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/support/thread/pthreads.hpp
@@ -130,7 +130,7 @@ namespace Gecode { namespace Support {
   forceinline unsigned int
   Thread::npu(void) {
 #ifdef GECODE_HAS_UNISTD_H
-    int n=sysconf(_SC_NPROCESSORS_ONLN);
+    int n=static_cast<int>(sysconf(_SC_NPROCESSORS_ONLN));
     return (n>1) ? n : 1;
 #else
     return 1;
